refactor(dashboard): extract tools-prompts domain (A1 phase 2 cont.)
- dashboard-tools-prompts.ts: ToolMetricsResponse/DashboardToolsResponse/
  DashboardScheduledAutomation* types + DashboardRuntimeProbe* types +
  PromptSource/DashboardPromptItem + fetchToolMetrics/fetchDashboardRuntimeProbe/
  fetchDashboardTools/fetchDashboardPrompts/save+clearPromptOverride.
  Tools/prompts/probe blocks were split around a ../types re-export (preserved).
- dashboard.ts: facade re-export only; consumers unchanged
- ripple import cleanup: TelemetryFreshnessMetadata, DashboardConfigResolution,
  DashboardRuntimeResolution (all moved to domain file or already re-exported
  via direct from)
- typecheck green (tsc --noEmit)

dashboard.ts 1888 -> 1542 lines after tools-prompts.
