+++ dashboard/src/api/dashboard.ts
@@ -4,7 +4,6 @@ import { isRecord, asBoolean, asInt, asNullableString, asNumber, asRecordArray, 
 import { normalizeKeeperTrustTerminalReason } from '../keeper-store-normalize'
 import { get, post, type AbortableRequestOptions } from './core'
 import { ensureDevToken } from './dev-token'
-import type { TelemetryFreshnessMetadata } from './dashboard-shared'
 import {
   type AgentTimelineEvent,
   type AgentTimelineResponse,
@@ -44,8 +43,6 @@ import type {
   GoalVerificationRequest,
   GoalVerificationSummary,
   GoalVerificationVote,
-  DashboardConfigResolution,
-  DashboardRuntimeResolution,
 } from '../types'
 export { DashboardConfigSchemaDriftError } from './schemas/dashboard-config'
 export type {
@@ -699,184 +696,20 @@ export async function fetchDashboardGoalDetail(goalId: string): Promise<Dashboar
   return decoded
 }
 
-// --- Tool metrics (P4 Phase 4.5) ---
-
-export interface DashboardToolInventoryItem {
-  name: string
-  description: string
-  category: string
-  category_description?: string | null
-  enabled_in_current_mode: boolean
-  direct_call_allowed: boolean
-  required_permission?: string | null
-  doc_refs: string[]
-  prompt_hints: string[]
-  surfaces: string[]
-  visibility: string
-  lifecycle: string
-  implementationStatus: string
-  tier: string
-  canonicalName?: string | null
-  replacement?: string | null
-  reason?: string | null
-}
-
-interface SurfaceSummaryEntry {
-  count: number
-  tools: string[]
-}
-
-interface DashboardToolInventoryResponse {
-  count: number
-  tools: DashboardToolInventoryItem[]
-  surface_summary?: Record<string, SurfaceSummaryEntry>
-}
-
-export interface ToolMetricsTopEntry {
-  name: string
-  call_count: number
-}
-
-export interface ToolMetricsResponse extends TelemetryFreshnessMetadata {
-  total_calls: number
-  distinct_tools_called: number
-  top_20: ToolMetricsTopEntry[]
-  never_called_count: number
-  tool_distribution?: { total: number; public: number; visible: number; hidden: number } | null
-  dispatch_v2_enabled: boolean
-  registered_count: number
-}
-
-export interface DashboardScheduledAutomationFsm {
-  state: string
-  active_count: number
-  terminal_count: number
-  next_due_at?: string | null
-}
-
-export interface DashboardScheduledAutomationExecution {
-  execution_id: string
-  schedule_id: string
-  started_at?: number
-  started_at_iso?: string | null
-  finished_at?: number | null
-  finished_at_iso?: string | null
-  due_at?: number
-  payload_digest?: string
-  status: string
-  detail?: unknown | null
-  error?: string | null
-}
-
-export interface DashboardScheduledAutomationKeeperToolStatus {
-  name: string
-  registered_schema?: boolean
-  dispatch_registered?: boolean
-  direct_call_allowed?: boolean
-  visibility?: string
-  surfaces?: string[]
-  surface_count?: number
-  effect_domain?: string | null
-  read_only?: boolean | null
-  requires_actor_binding?: boolean | null
-}
-
-export interface DashboardScheduledAutomationActor {
-  id: string
-  kind: string
-  display_name?: string | null
-}
-
-export interface DashboardScheduledAutomationSignal {
-  signal_id: string
-  kind: string
-  event_type?: string
-  schedule_id: string
-  emitted_at?: number
-  emitted_at_iso?: string | null
-  due_at?: number
-  due_at_iso?: string | null
-  risk_class: string
-  payload_digest?: string
-  payload_kind?: string | null
-}
-
-export interface DashboardScheduledAutomationRequest {
-  schedule_id: string
-  status: string
-  effective_status?: string
-  execution_readiness?: string
-  operator_action?: string | null
-  keeper_next_tool?: string | null
-  keeper_next_tool_status?: DashboardScheduledAutomationKeeperToolStatus | null
-  keeper_next_action?: string | null
-  risk_class: string
-  approval_required: boolean
-  source: string
-  requested_by?: DashboardScheduledAutomationActor | null
-  scheduled_by?: DashboardScheduledAutomationActor | null
-  recurrence?: {
-    kind: string
-    interval_sec?: number
-    hour?: number
-    minute?: number
-    second?: number
-    expression?: string
-    timezone?: string
-  }
-  recurrence_kind?: string
-  requested_at?: number
-  requested_at_iso?: string
-  due_at?: number
-  due_at_iso?: string
-  next_due_at?: number | null
-  next_due_at_iso?: string | null
-  expires_at?: number | null
-  expires_at_iso?: string | null
-  payload_digest?: string
-  payload_kind?: string | null
-  payload_support?: 'supported' | 'unsupported' | 'unknown'
-  payload_target?: string | null
-  payload_summary?: string | null
-  recurrence_summary?: string | null
-  requires_separate_human_grant?: boolean
-  approval_policy?: string | null
-  last_execution?: DashboardScheduledAutomationExecution | null
-}
-
-export interface DashboardScheduledAutomationPayloadSupport {
-  supported_kinds?: string[]
-  unsupported_request_count?: number
-  unsupported_kinds?: Array<{ kind: string; count: number }>
-  unknown_request_count?: number
-}
-
-export interface DashboardScheduledAutomation {
-  schema?: string
-  source?: string
-  generated_at?: string
-  request_count: number
-  request_limit: number
-  truncated: boolean
-  signal_source?: string
-  signal_count?: number
-  signal_limit?: number
-  signals?: DashboardScheduledAutomationSignal[]
-  counts: Record<string, number>
-  derived_counts?: Record<string, number>
-  payload_support?: DashboardScheduledAutomationPayloadSupport
-  fsm: DashboardScheduledAutomationFsm
-  requests: DashboardScheduledAutomationRequest[]
-}
-
-export interface DashboardToolsResponse {
-  generated_at?: string
-  config_resolution?: DashboardConfigResolution
-  runtime_resolution?: DashboardRuntimeResolution
-  tool_inventory: DashboardToolInventoryResponse
-  tool_usage: ToolMetricsResponse
-  scheduled_automation?: DashboardScheduledAutomation
-}
+export type {
+  DashboardToolInventoryItem,
+  ToolMetricsTopEntry,
+  ToolMetricsResponse,
+  DashboardScheduledAutomationFsm,
+  DashboardScheduledAutomationExecution,
+  DashboardScheduledAutomationKeeperToolStatus,
+  DashboardScheduledAutomationActor,
+  DashboardScheduledAutomationSignal,
+  DashboardScheduledAutomationRequest,
+  DashboardScheduledAutomationPayloadSupport,
+  DashboardScheduledAutomation,
+  DashboardToolsResponse,
+} from './dashboard-tools-prompts'
 
 export type {
   DashboardConfigResolution,
@@ -891,204 +724,22 @@ export type {
   KeeperRuntimeSource,
 } from '../types'
 
-interface DashboardRuntimeProbeLoadedModel {
-  name?: string | null
-  model?: string | null
-  size_vram_bytes?: number | null
-  context_length?: number | null
-  expires_at?: string | null
-}
-
-interface DashboardRuntimeProbeRun {
-  run_index: number
-  http_status?: number | null
-  wall_clock_ms?: number | null
-  total_duration_ms?: number | null
-  load_duration_ms?: number | null
-  prompt_eval_count?: number | null
-  prompt_eval_duration_ms?: number | null
-  prompt_tokens_per_second?: number | null
-  eval_count?: number | null
-  eval_duration_ms?: number | null
-  generation_tokens_per_second?: number | null
-  done?: boolean | null
-  done_reason?: string | null
-  thinking_present?: boolean
-  response_preview?: string | null
-  response_chars?: number | null
-  error?: string | null
-}
-
-interface DashboardRuntimeProbeAssessment {
-  signal?: string | null
-  baseline_run_index?: number | null
-  best_repeat_run_index?: number | null
-  baseline_prompt_eval_duration_ms?: number | null
-  best_repeat_prompt_eval_duration_ms?: number | null
-  prompt_eval_duration_reduction_ratio?: number | null
-  note?: string | null
-  limitation?: string | null
-}
-
-export interface DashboardRuntimeProviderProbe {
-  runtime_id?: string | null
-  provider_id?: string | null
-  provider_display_name?: string | null
-  model_id?: string | null
-  model_api_name?: string | null
-  protocol?: string | null
-  runtime_kind?: string | null
-  transport?: string | null
-  auth_kind?: string | null
-  credential_required?: boolean | null
-  auth_present?: boolean | null
-  status?: string | null
-  reachable?: boolean | null
-  http_status?: number | null
-  latency_ms?: number | null
-  model_count?: number | null
-  content_type?: string | null
-  downloaded_bytes?: number | null
-  endpoint_url?: string | null
-  probe_url?: string | null
-  error?: string | null
-  checked_at?: string | null
-}
-
-export interface DashboardRuntimeProviderProbeSummary {
-  runtimes?: number
-  probed?: number
-  reachable?: number
-  failed?: number
-  skipped?: number
-  default_runtime_id?: string | null
-}
-
-export interface DashboardRuntimeProbePayload {
-  source?: string
-  status?: string | null
-  checked_at?: string | null
-  summary?: DashboardRuntimeProviderProbeSummary | null
-  providers?: DashboardRuntimeProviderProbe[]
-  server_url?: string
-  ps_endpoint?: string
-  generate_endpoint?: string
-  configured_default_model?: string | null
-  requested_model?: string | null
-  effective_model?: string | null
-  probe_runs_requested?: number
-  probe_runs_completed?: number
-  max_tokens?: number
-  keep_alive?: string | null
-  timeout_sec?: number
-  ps_timeout_sec?: number
-  prompt_chars?: number
-  prompt_preview?: string
-  ps_http_status_before?: number | null
-  ps_http_status_after?: number | null
-  loaded_models_before?: DashboardRuntimeProbeLoadedModel[]
-  loaded_models_after?: DashboardRuntimeProbeLoadedModel[]
-  model_loaded_before_probe?: boolean
-  model_loaded_after_probe?: boolean
-  runs?: DashboardRuntimeProbeRun[]
-  kv_cache_assessment?: DashboardRuntimeProbeAssessment | null
-  observations?: string[]
-  errors?: string[]
-  limitations?: string[]
-  probe_ok?: boolean
-}
-
-export interface DashboardRuntimeProbeResponse {
-  generated_at?: string
-  refreshed_at_unix?: number
-  cache_ttl_sec?: number
-  cache_age_sec?: number
-  cache_hit?: boolean
-  // Non-blocking route freshness tag. 'served_stale' / 'warming_up' mean a
-  // background refresh was scheduled and the fresh value arrives on the next
-  // poll — a force=1 ("Live probe") response is not guaranteed to be fresh.
-  refresh_state?: 'fresh' | 'recent' | 'served_stale' | 'warming_up'
-  probe?: DashboardRuntimeProbePayload | null
-}
-
-export function fetchToolMetrics(): Promise<ToolMetricsResponse> {
-  return get('/api/v1/tool-metrics')
-}
-
-export async function fetchDashboardRuntimeProbe(
-  force = false,
-  opts?: AbortableRequestOptions,
-): Promise<DashboardRuntimeProbeResponse> {
-  const query = force ? '?force=1' : ''
-  await ensureDevToken()
-  return get(`/api/v1/dashboard/runtime-probe${query}`, { signal: opts?.signal })
-}
-
-export async function fetchDashboardTools(opts?: AbortableRequestOptions): Promise<DashboardToolsResponse> {
-  const raw = await get<DashboardToolsResponse>('/api/v1/dashboard/tools', { signal: opts?.signal })
-  const normalizedTools = raw.tool_inventory?.tools?.map(t => ({
-    ...t,
-    category: t.category ?? 'uncategorized',
-    tier: t.tier ?? '(unknown tier)',
-    // Tool-layer decoupling groundwork: surface membership is consumer-owned
-    // metadata, not an execution constraint. Totalize here so the field is
-    // never absent downstream; consumers keep working with [] and the surface
-    // filter simply degrades to zero counts. Mirrors category/tier above.
-    surfaces: t.surfaces ?? [],
-  }))
-  return {
-    ...raw,
-    tool_inventory: {
-      ...raw.tool_inventory,
-      ...(normalizedTools ? { tools: normalizedTools } : {}),
-    },
-  }
-}
-
-export type PromptSource = 'override' | 'file' | 'default' | 'missing'
-
-export interface DashboardPromptItem {
-  key: string
-  category: string
-  description: string
-  current: string
-  default: string | null
-  effective: string
-  file_value: string | null
-  override_value: string | null
-  file_path: string | null
-  file_exists: boolean
-  source: PromptSource
-  has_override: boolean
-  char_count: number
-  required_file: boolean
-  template_variables: string[]
-}
-
-interface DashboardPromptsResponse {
-  prompts: DashboardPromptItem[]
-}
-
-interface PromptMutationResponse {
-  ok: boolean
-  message?: string
-  key?: string
-  source?: PromptSource
-  effective?: string
-  error?: string
-}
-
-export function fetchDashboardPrompts(): Promise<DashboardPromptsResponse> {
-  return get('/api/v1/prompts')
-}
-
-export function savePromptOverride(key: string, value: string): Promise<PromptMutationResponse> {
-  return post('/api/v1/prompts', { action: 'set', key, value })
-}
-
-export function clearPromptOverride(key: string): Promise<PromptMutationResponse> {
-  return post('/api/v1/prompts', { action: 'clear', key })
-}
+export type {
+  DashboardRuntimeProviderProbe,
+  DashboardRuntimeProviderProbeSummary,
+  DashboardRuntimeProbePayload,
+  DashboardRuntimeProbeResponse,
+  PromptSource,
+  DashboardPromptItem,
+} from './dashboard-tools-prompts'
+export {
+  fetchToolMetrics,
+  fetchDashboardRuntimeProbe,
+  fetchDashboardTools,
+  fetchDashboardPrompts,
+  savePromptOverride,
+  clearPromptOverride,
+} from './dashboard-tools-prompts'
 
 function asLooseBoolean(value: unknown, fallback = false): boolean {
   const booleanValue = asBoolean(value)
